+++ packages/riverpod/lib/src/provider/base.dart
@@ -2,7 +2,7 @@ part of '../provider.dart';
 
 /// {@macro riverpod.providerrefbase}
 /// - [state], the value currently exposed by this providers.
-abstract class ProviderRef<State> implements ProviderRefBase {
+abstract class ProviderRef<State> implements Ref {
   /// Obtains the state currently exposed by this provider.
   ///
   /// Mutating this property will notify the provider listeners.
@@ -175,7 +175,7 @@ abstract class ProviderRef<State> implements ProviderRefBase {
 /// class Location {
 ///   Location(this._ref);
 ///
-///   final ProviderRefBase _ref;
+///   final Ref _ref;
 ///
 ///   String get label {
 ///     final city = _ref.read(cityProvider);