+++ packages/riverpod/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Breaking** `AsyncValue.value` now return `null` instead of throwing if
   used during loading state (#270)
+- **Breaking** Renamd `ProviderRefBase` to `Ref`.
 - Added an assertion that detects when a provider override is missing.
 - Added an assert for when a provider specified `dependencies` but is
   depending on a provider that isn't in the list.